Set application home URI to /dashboard (#181)
Set the application home to /dashboard since the /home URI does not exist by default but /dashboard exists on the Breeze Next.js example app. Alternatively, this could be set to "/" as this URL should be available on most apps. This then will be used by the VerifyEmailController for redirection to config('app.frontend_url').RouteServiceProvider::HOME.'?verified=1'.